refactor TxDescription component
